layout/scrolling: make mouse click viewport move consistent across all `follow_mouse` options (#15975)
* CLICK_DOWN (mouser-down) no longer allowed to move viewport in scrolling - it still moves forcus just fine, just not the viewport
Add 2 new focus reasons - click up and down. Made the enum a bitmap. Allows users to have more info on focus change reason view `window.active` event

* woops

* not here

* TEST

* format

* Ensure op Precedence

+++ src/desktop/view/window/Window.hpp
@@ -33,7 +33,7 @@ namespace Layout {
 }
 
 namespace Desktop {
-    enum eFocusReason : uint8_t;
+    enum eFocusReason : uint32_t;
 }
 
 namespace Desktop::View {
